+++ proton-mail-uniffi/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased] - 2024-00-00
 
+## [0.11.27] - 2024-09-02
+
 ### Added
 
   - `apply_label_to_messages` who apply a label to many messages.